+++ main.go
@@ -11,8 +11,8 @@ func main() {
 	defer window.Destroy()
 
 	engine := NewEngine(window, renderer)
-	// engine.LoadCubeMesh()
-	engine.LoadMesh("assets/f22.obj")
+	engine.LoadCubeMesh()
+	// engine.LoadMesh("assets/f22.obj")
 	engine.Setup()
 
 	for engine.isRunning {