+++ texture.go
@@ -1,0 +1,6 @@
+package main
+
+type Texture struct {
+	width, height int
+	data          []Color
+}