+++ triangle.go
@@ -1,14 +1,9 @@
 package main
 
-// Face represents a triangle before rasterization.
-type Face struct {
-	points [3]Vec3
-	color  Color
-}
-
 // Triangle represents a triangle after rasterization.
 type Triangle struct {
 	points       [3]Vec4
+	texcoords    [3]Tex
 	color        Color
 	averageDepth float64
 }