+++ HISTORY.md
@@ -66,6 +66,7 @@ Hypotenous / |
 
 - [x] Add perspective projection matrix
 - [x] Add flat light shading
+- [x] Add texture mapping
 
 ##### Notes: